backend/requirements.txt
flask==3.0.3
influxdb-client==1.44.0
python-dotenv==1.0.1